Minor fixes and core update
