+++ EmmetNPP/emmet/pyv8loader.py
@@ -8,6 +8,10 @@ import threading
 import subprocess
 import tempfile
 import collections
+import platform
+import semver
+import time
+import zipfile
 
 is_python3 = sys.version_info[0] > 2
 
@@ -22,8 +26,175 @@ else:
 	url_err = urllib2
 	url_parse = urllib2
 
+CHECK_INTERVAL = 60 * 60 * 24
+
 # PACKAGES_URL = 'https://api.github.com/repos/emmetio/pyv8-binaries/downloads'
 PACKAGES_URL = 'https://api.github.com/repos/emmetio/pyv8-binaries/contents'
+
+def load(dest_path, delegate=None):
+	"""
+	Main function that attempts to load or update PyV8 binary.
+	First, it loads list of available PyV8 modules and check if
+	PyV8 should be downloaded or updated.
+	@param dest_path: Path where PyV8 lib should be downloaded 
+	@param delegate: instance of LoaderDelegate that will receive
+	loader progress events
+	@returns: `True` if download progress was initiated
+	"""
+	if delegate is None:
+		delegate = LoaderDelegate()
+
+	config = get_loader_config(dest_path)
+
+	if 'PyV8' in sys.modules and (config['skip_update'] or time.time() < config['last_update'] + CHECK_INTERVAL):
+		# No need to load anything: user already has PyV8 binary
+		# or decided to disable update process
+		delegate.log('No need to update PyV8')
+		return False
+
+	def on_complete(result, thread):
+		if result is not None:
+			# Most recent version was downloaded
+			config['last_id'] = result				
+			if 'PyV8' not in sys.modules:
+				# PyV8 is not loaded yet, we can safely unpack it 
+				unpack_pyv8(dest_path)
+
+		config['last_update'] = time.time()
+		save_loader_config(dest_path, config)
+		delegate.on_complete()
+
+	# try to download most recent version of PyV8
+	thread = PyV8Loader(get_arch(), dest_path, config, delegate=delegate)
+	thread.start()
+	delegate.on_start()
+	
+	# watch on download progress
+	prog = ThreadProgress(thread, delegate)
+	prog.on('complete', on_complete)
+	prog.on('error', delegate.on_error)
+
+def get_arch():
+	"Returns architecture name for PyV8 binary"
+	suffix = is_python3 and '-p3' or ''
+	p = lambda a: '%s%s' % (a, suffix)
+	is_64bit = sys.maxsize > 2**32
+	system_name = platform.system()
+	if system_name == 'Darwin':
+		if semver.match(platform.mac_ver()[0], '<10.7.0'):
+			return p('mac106')
+
+		return p('osx')
+	if system_name == 'Windows':
+		return p('win64') if is_64bit else p('win32')
+	if system_name == 'Linux':
+		return p('linux64') if is_64bit else p('linux32')
+
+def get_loader_config(path):
+	config = {
+		"last_id": 0,
+		"last_update": 0,
+		"skip_update": False
+	}
+
+	config_path = os.path.join(path, 'config.json')
+	if os.path.exists(config_path):
+		with open(config_path) as fd:
+			for k,v in json.load(fd).items():
+				config[k] = v
+
+	return config
+
+def save_loader_config(path, data):
+	config_path = os.path.join(path, 'config.json')
+	
+	if not os.path.exists(path):
+		os.makedirs(path)
+	fp = open(config_path, 'w')
+	fp.write(json.dumps(data))
+	fp.close()
+
+def unpack_pyv8(package_dir):
+	f = os.path.join(package_dir, 'pack.zip')
+	if not os.path.exists(f):
+		return
+
+	package_zip = zipfile.ZipFile(f, 'r')
+
+	root_level_paths = []
+	last_path = None
+	for path in package_zip.namelist():
+		last_path = path
+		if path.find('/') in [len(path) - 1, -1]:
+			root_level_paths.append(path)
+		if path[0] == '/' or path.find('../') != -1 or path.find('..\\') != -1:
+			raise 'The PyV8 package contains files outside of the package dir and cannot be safely installed.'
+
+	if last_path and len(root_level_paths) == 0:
+		root_level_paths.append(last_path[0:last_path.find('/') + 1])
+
+	prev_dir = os.getcwd()
+	os.chdir(package_dir)
+
+	# Here we don't use .extractall() since it was having issues on OS X
+	skip_root_dir = len(root_level_paths) == 1 and \
+		root_level_paths[0].endswith('/')
+	extracted_paths = []
+	for path in package_zip.namelist():
+		dest = path
+
+		if not is_python3:
+			try:
+				if not isinstance(dest, unicode):
+					dest = unicode(dest, 'utf-8', 'strict')
+			except UnicodeDecodeError:
+				dest = unicode(dest, 'cp1252', 'replace')
+
+		if os.name == 'nt':
+			regex = ':|\*|\?|"|<|>|\|'
+			if re.search(regex, dest) != None:
+				print ('%s: Skipping file from package named %s due to ' +
+					'an invalid filename') % (__name__, path)
+				continue
+
+		# If there was only a single directory in the package, we remove
+		# that folder name from the paths as we extract entries
+		if skip_root_dir:
+			dest = dest[len(root_level_paths[0]):]
+
+		if os.name == 'nt':
+			dest = dest.replace('/', '\\')
+		else:
+			dest = dest.replace('\\', '/')
+
+		dest = os.path.join(package_dir, dest)
+
+		def add_extracted_dirs(dir):
+			while dir not in extracted_paths:
+				extracted_paths.append(dir)
+				dir = os.path.dirname(dir)
+				if dir == package_dir:
+					break
+
+		if path.endswith('/'):
+			if not os.path.exists(dest):
+				os.makedirs(dest)
+			add_extracted_dirs(dest)
+		else:
+			dest_dir = os.path.dirname(dest)
+			if not os.path.exists(dest_dir):
+				os.makedirs(dest_dir)
+			add_extracted_dirs(dest_dir)
+			extracted_paths.append(dest)
+			try:
+				open(dest, 'wb').write(package_zip.read(path))
+			except (IOError, UnicodeDecodeError):
+				print ('%s: Skipping file from package named %s due to ' +
+					'an invalid filename') % (__name__, path)
+	package_zip.close()
+
+	os.chdir(prev_dir)
+	os.remove(f)
 
 class LoaderDelegate():
 	"""
@@ -52,6 +223,9 @@ class LoaderDelegate():
 	def setting(self, name, default=None):
 		"Returns specified setting name"
 		return self.settings[name] if name in self.settings else default
+
+	def log(self, message):
+		pass
 
 class ThreadProgress():
 	def __init__(self, thread, delegate):
@@ -294,19 +468,16 @@ class UrlLib2Downloader():
 		return False
 
 class PyV8Loader(threading.Thread):
-	def __init__(self, arch, download_path, config):
+	def __init__(self, arch, download_path, config, delegate=None):
 		self.arch = arch
 		self.config = config
 		self.download_path = download_path
 		self.exit_code = 0
 		self.result = None
-		self.settings = {}
+		self.delegate = delegate or LoaderDelegate()
 
 		threading.Thread.__init__(self)
-		self.log('Creating thread')
-
-	def log(self, message):
-		print('PyV8 Loader: %s' % message)
+		self.delegate.log('Creating thread')
 
 	def download_url(self, url, error_message):
 		# TODO add settings
@@ -314,17 +485,17 @@ class PyV8Loader(threading.Thread):
 		is_ssl = re.search('^https://', url) != None
 
 		if (is_ssl and has_ssl) or not is_ssl:
-			downloader = UrlLib2Downloader(self.settings)
+			downloader = UrlLib2Downloader(self.delegate.settings)
 		else:
 			for downloader_class in [CurlDownloader, WgetDownloader]:
 				try:
-					downloader = downloader_class(self.settings)
+					downloader = downloader_class(self.delegate.settings)
 					break
 				except BinaryNotFoundError:
 					pass
 
 		if not downloader:
-			self.log('Unable to download PyV8 binary due to invalid downloader')
+			self.delegate.log('Unable to download PyV8 binary due to invalid downloader')
 			return False
 
 		# timeout = self.settings.get('timeout', 3)
@@ -333,8 +504,13 @@ class PyV8Loader(threading.Thread):
 
 	def run(self):
 		# get list of available packages first
-		self.log('Loading %s' % PACKAGES_URL)
-		packages = self.download_url(PACKAGES_URL, 'Unable to download packages list.')
+		self.delegate.log('Loading %s' % PACKAGES_URL)
+		try:
+			packages = self.download_url(PACKAGES_URL, 'Unable to download packages list.')
+		except Exception as e:
+			self.delegate.log('Unable to download file: %s' % e)
+			self.exit_code = 4
+			return
 
 		if not packages:
 			self.exit_code = 1
@@ -354,16 +530,16 @@ class PyV8Loader(threading.Thread):
 				break
 
 		if not cur_item:
-			self.log('Unable to find binary for %s architecture' % self.arch)
+			self.delegate.log('Unable to find binary for %s architecture' % self.arch)
 			self.exit_code = 2
 			return
 
 		if cur_item['sha'] == self.config['last_id']:
-			self.log('You have the most recent PyV8 binary')
+			self.delegate.log('You have the most recent PyV8 binary')
 			return
 
 		url = 'https://raw.github.com/emmetio/pyv8-binaries/master/%s' % cur_item['name']
-		self.log('Loading PyV8 binary from %s' % url)
+		self.delegate.log('Loading PyV8 binary from %s' % url)
 		package = self.download_url(url, 'Unable to download package from %s' % url)
 		if not package:
 			self.exit_code = 3
